More refactoring
