+++ public/app/features/dashboard/components/Inspector/InspectContent.tsx
@@ -1,10 +1,8 @@
 import React, { useState } from 'react';
-import { DataSourceApi, PanelData, PanelPlugin } from '@grafana/data';
+import { DataSourceApi, formattedValueToString, getValueFormat, PanelData, PanelPlugin } from '@grafana/data';
 import { getTemplateSrv } from '@grafana/runtime';
-import { CustomScrollbar, Drawer, TabContent } from '@grafana/ui';
-import { getPanelInspectorStyles } from 'app/features/inspector/styles';
+import { Drawer, Tab, TabsBar } from '@grafana/ui';
 import { InspectMetadataTab } from 'app/features/inspector/InspectMetadataTab';
-import { InspectSubtitle } from 'app/features/inspector/InspectSubtitle';
 import { InspectJSONTab } from 'app/features/inspector/InspectJSONTab';
 import { QueryInspector } from 'app/features/inspector/QueryInspector';
 import { InspectStatsTab } from 'app/features/inspector/InspectStatsTab';
@@ -13,7 +11,6 @@ import { InspectDataTab } from 'app/features/inspector/InspectDataTab';
 import { InspectTab } from 'app/features/inspector/types';
 import { DashboardModel, PanelModel } from '../../state';
 import { GetDataOptions } from '../../../query/state/PanelQueryRunner';
-import { InspectActionsTab } from './PanelInspectActions';
 
 interface Props {
   dashboard: DashboardModel;
@@ -50,7 +47,6 @@ export const InspectContent: React.FC<Props> = ({
     return null;
   }
 
-  const styles = getPanelInspectorStyles();
   const error = data?.error;
 
   // Validate that the active tab is actually valid and allowed
@@ -58,22 +54,31 @@ export const InspectContent: React.FC<Props> = ({
   if (!tabs.find((item) => item.value === currentTab)) {
     activeTab = InspectTab.JSON;
   }
+
   const title = getTemplateSrv().replace(panel.title, panel.scopedVars, 'text');
 
   return (
     <Drawer
       title={`Inspect: ${title || 'Panel'}`}
-      subtitle={
-        <InspectSubtitle
-          tabs={tabs}
-          tab={activeTab}
-          data={data}
-          onSelectTab={(item) => setCurrentTab(item.value || InspectTab.Data)}
-        />
-      }
+      subtitle={data && formatStats(data)}
       width="50%"
       onClose={onClose}
       expandable
+      scrollableContent
+      tabs={
+        <TabsBar>
+          {tabs.map((t, index) => {
+            return (
+              <Tab
+                key={`${t.value}-${index}`}
+                label={t.label}
+                active={t.value === activeTab}
+                onChangeTab={() => setCurrentTab(t.value || InspectTab.Data)}
+              />
+            );
+          })}
+        </TabsBar>
+      }
     >
       {activeTab === InspectTab.Data && (
         <InspectDataTab
@@ -85,23 +90,31 @@ export const InspectContent: React.FC<Props> = ({
           timeZone={dashboard.timezone}
         />
       )}
-      <CustomScrollbar autoHeightMin="100%">
-        <TabContent className={styles.tabContent}>
-          {data && activeTab === InspectTab.Meta && (
-            <InspectMetadataTab data={data} metadataDatasource={metadataDatasource} />
-          )}
+      {data && activeTab === InspectTab.Meta && (
+        <InspectMetadataTab data={data} metadataDatasource={metadataDatasource} />
+      )}
 
-          {activeTab === InspectTab.JSON && (
-            <InspectJSONTab panel={panel} dashboard={dashboard} data={data} onClose={onClose} />
-          )}
-          {activeTab === InspectTab.Error && <InspectErrorTab error={error} />}
-          {data && activeTab === InspectTab.Stats && <InspectStatsTab data={data} timeZone={dashboard.getTimezone()} />}
-          {data && activeTab === InspectTab.Query && (
-            <QueryInspector panel={panel} data={data.series} onRefreshQuery={() => panel.refresh()} />
-          )}
-          {activeTab === InspectTab.Actions && <InspectActionsTab panel={panel} data={data} />}
-        </TabContent>
-      </CustomScrollbar>
+      {activeTab === InspectTab.JSON && (
+        <InspectJSONTab panel={panel} dashboard={dashboard} data={data} onClose={onClose} />
+      )}
+      {activeTab === InspectTab.Error && <InspectErrorTab error={error} />}
+      {data && activeTab === InspectTab.Stats && <InspectStatsTab data={data} timeZone={dashboard.getTimezone()} />}
+      {data && activeTab === InspectTab.Query && (
+        <QueryInspector panel={panel} data={data.series} onRefreshQuery={() => panel.refresh()} />
+      )}
     </Drawer>
   );
 };
+
+function formatStats(data: PanelData) {
+  const { request } = data;
+  if (!request) {
+    return '';
+  }
+
+  const queryCount = request.targets.length;
+  const requestTime = request.endTime ? request.endTime - request.startTime : 0;
+  const formatted = formattedValueToString(getValueFormat('ms')(requestTime));
+
+  return `${queryCount} queries with total query time of ${formatted}`;
+}
