+++ public/app/features/inspector/InspectDataTab.tsx
@@ -14,7 +14,7 @@ import {
   transformDataFrame,
   TimeZone,
 } from '@grafana/data';
-import { Button, Container, Spinner, Table } from '@grafana/ui';
+import { Button, Spinner, Table } from '@grafana/ui';
 import { selectors } from '@grafana/e2e-selectors';
 import { InspectDataOptions } from './InspectDataOptions';
 import { getPanelInspectorStyles } from './styles';
@@ -232,8 +232,8 @@ export class InspectDataTab extends PureComponent<Props, State> {
     const hasTraces = dataFrames.some((df) => df?.meta?.preferredVisualisationType === 'trace');
 
     return (
-      <div className={styles.dataTabContent} aria-label={selectors.components.PanelInspector.Data.content}>
-        <div className={styles.actionsWrapper}>
+      <div className={styles.wrap} aria-label={selectors.components.PanelInspector.Data.content}>
+        <div className={styles.toolbar}>
           <InspectDataOptions
             data={data}
             panel={panel}
@@ -281,7 +281,7 @@ export class InspectDataTab extends PureComponent<Props, State> {
             </Button>
           )}
         </div>
-        <Container grow={1}>
+        <div className={styles.content}>
           <AutoSizer>
             {({ width, height }) => {
               if (width === 0) {
@@ -295,7 +295,7 @@ export class InspectDataTab extends PureComponent<Props, State> {
               );
             }}
           </AutoSizer>
-        </Container>
+        </div>
       </div>
     );
   }
