+++ public/app/features/inspector/InspectJSONTab.tsx
@@ -129,7 +129,7 @@ export class InspectJSONTab extends PureComponent<Props, State> {
     const styles = getPanelInspectorStyles();
 
     return (
-      <>
+      <div className={styles.wrap}>
         <div className={styles.toolbar} aria-label={selectors.components.PanelInspector.Json.content}>
           <Field label="Select source" className="flex-grow-1">
             <Select
@@ -162,7 +162,7 @@ export class InspectJSONTab extends PureComponent<Props, State> {
             )}
           </AutoSizer>
         </div>
-      </>
+      </div>
     );
   }
 }
