+++ public/app/features/inspector/styles.ts
@@ -25,11 +25,6 @@ export const getPanelInspectorStyles = stylesFactory(() => {
     content: css`
       flex-grow: 1;
       height: 100%;
-      padding-bottom: 16px;
-    `,
-    contentQueryInspector: css`
-      flex-grow: 1;
-      padding: ${config.theme.spacing.md} 0;
     `,
     editor: css`
       font-family: monospace;
@@ -41,20 +36,6 @@ export const getPanelInspectorStyles = stylesFactory(() => {
     `,
     dataFrameSelect: css`
       flex-grow: 2;
-    `,
-    tabContent: css`
-      height: 100%;
-      display: flex;
-      flex-direction: column;
-    `,
-    dataTabContent: css`
-      display: flex;
-      flex-direction: column;
-      height: 100%;
-      width: 100%;
-    `,
-    actionsWrapper: css`
-      display: flex;
     `,
     leftActions: css`
       display: flex;
